Iterable abstract class interface

An Iterable is an object that provides an iter() method, which is expected to return an Iterator object. This class defines the abstract class interface such that each class inheriting this class provides an iter() method and must implement a private method create_iter(), which must return an Iterator object.

See also

Author

Roman Pahl

Methods

Public methods

Iterable$new()

Iterable$iter()

Iterable$clone()

Method new()

Iterable is an abstract class and thus cannot be instantiated.

Usage

Iterable$new()

Method iter()

Create iterator

Usage

Iterable$iter()

Returns

returns the Iterator object.

Method clone()

The objects of this class are cloneable with this method.

Usage

Iterable$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.