OrderedSet Class

The OrderedSet is a Set where all elements are always ordered.

Details

The order of elements is determined sequentially as follows:

element's length

whether it is an atomic element

the element's class(es)

by numeric value (if applicable)

it's representation when printed

the name of the element in the Set

See also

Super classes

container::Iterable -> container::Container -> container::Set -> OrderedSet

Methods

Public methods

OrderedSet$new()

OrderedSet$add()

OrderedSet$clone()

Inherited methods

container::Iterable$iter()

container::Container$at()

container::Container$at2()

container::Container$clear()

container::Container$count()

container::Container$delete()

container::Container$delete_at()

container::Container$discard()

container::Container$discard_at()

container::Container$empty()

container::Container$get_compare_fun()

container::Container$has()

container::Container$has_name()

container::Container$is_empty()

container::Container$length()

container::Container$names()

container::Container$peek_at()

container::Container$peek_at2()

container::Container$pop()

container::Container$print()

container::Container$remove()

container::Container$rename()

container::Container$replace()

container::Container$replace_at()

container::Container$size()

container::Container$type()

container::Container$update()

container::Set$diff()

container::Set$intersect()

container::Set$is_equal()

container::Set$is_proper_subset()

container::Set$is_subset()

container::Set$union()

container::Set$values()

Method new()

OrderedSet constructor

Usage

OrderedSet$new(...)

Arguments

...

initial elements put into the OrderedSet

Returns

returns the OrderedSet object

Method add()

Add element

Usage

OrderedSet$add(value, name = NULL)

Arguments

value

value of ANY type to be added to the OrderedSet.

name

character optional name attribute of the value.

Returns

the OrderedSet object.

Method clone()

The objects of this class are cloneable with this method.

Usage

OrderedSet$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

s1 = OrderedSet$new(2, 1)
s1

#> {1, 2}