Set Container Package Options

Set Container Package Options

container_options(..., .reset = FALSE)

getContainerOption(x, default = NULL)

Arguments

...

any options can be defined, using name = value.

.reset

logical if TRUE, the options are reset to their default and returned.

x

a character string holding an option name.

default

if the specified option is not set in the options list, this value is returned.

Value

container_options() returns a list of all set options sorted by name.

container_options(name), a list of length one containing the set value, or NULL if it is unset. Can also be multiple names (see Examples).

container_options(key = value) sets the option with name key to value and returns the previous options invisibly.

Container Options

compare (default = all.equal)

useDots (default = TRUE) whether to abbreviate long container elements with ... when exceeding vec.len (see below). If FALSE, they are abbreviated as <<type(length)>>.

vec.len (default = 4) the length limit at which container vectors are abbreviated.

Examples

co = container(1L, 1:10, as.list(1:5))
co
#> [1L, (1L 2L 3L 4L ...), <<list(5)>>]

container_options(useDots = FALSE)
co
#> [1L, <<integer(10)>>, <<list(5)>>]

container_options(useDots = TRUE, vec.len = 6)
co
#> [1L, (1L 2L 3L 4L 5L 6L ...), list(1L, 2L, 3L, 4L, 5L)]

has(co, 1.0)
#> [1] TRUE

container_options(compare = "identical")

has(co, 1.0) # still uses 'all.equal'
#> [1] TRUE

co2 = container(1L)
has(co2, 1.0)
#> [1] FALSE
has(co2, 1L)
#> [1] TRUE

container_options()
#> $compare
#> [1] "identical"
#> 
#> $useDots
#> [1] TRUE
#> 
#> $vec.len
#> [1] 6
#> 
container_options(.reset = TRUE)
#> $compare
#> [1] "all.equal"
#> 
#> $useDots
#> [1] TRUE
#> 
#> $vec.len
#> [1] 4
#>