Set and ordered Set

The Set is considered and implemented as a specialized Container, that is, Set elements are always unique. It provides typical set operations such as union and intersect.

setnew(..., .ordered = FALSE)

as.set(x)

as.orderedset(x)

is.set(x)

is.orderedset(x)

Arguments

...
initial elements put into the Set.

.ordered
logical if TRUE all elements in the Set will be ordered.

x
R object of ANY type for as.set() and is.set() or of class Set for the S3 methods.

Details

Methods that alter Set objects usually come in two versions providing either copy or reference semantics where the latter start with 'ref_' to note the reference semantic, for example, add() and ref_add().

setnew(...) initializes and returns a Set() object.

as.set(x) coerces x to a set.

as.orderedset(x) coerces x to an ordered set.

is.set(x) returns TRUE if x is of class Set and FALSE otherwise.

is.orderedset(x) returns TRUE if x is of class OrderedSet and FALSE otherwise.

x & y performs the set intersection of x and y

x | y performs the set union of x and y

See also

Examples

s = setnew(1, b = NA, 1:3, c = container("a", 1))
is.set(s)

#> [1] TRUE

print(s)

#> {1, b = NA, (1L 2L 3L), c = ["a", 1]}

length(s)

#> [1] 4

names(s)

#> [1] ""  "b" ""  "c"

as.list(s)

#> [[1]]
#> [1] 1
#> 
#> $b
#> [1] NA
#> 
#> [[3]]
#> [1] 1 2 3
#> 
#> $c
#> ["a", 1]
#> 

unpack(s)   # flatten recursively similar to unlist

#>       b              c1  c2 
#> "1"  NA "1" "2" "3" "a" "1" 

so = setnew(2, 1, .ordered = TRUE)
print(so)

#> {1, 2}

add(so, 0)

#> {0, 1, 2}

# Math
s = setnew(5:3, 1, 2)
s

#> {(5L 4L 3L), 1, 2}

abs(s)

#> {5, 4, 3, 1, 2}

cumsum(s)

#> {5, 9, 12, 13, 15}

round(s)

#> {5, 4, 3, 1, 2}

exp(s)

#> {148.4132, 54.59815, 20.08554, 2.718282, 7.389056}

# Summary
range(s)

#> [1] 1 5

min(s)

#> [1] 1

max(s)

#> [1] 5

s1 = setnew(1, 1:2)
s2 = setnew(2, 1:2)
s1 + s2     # same as s1 | s2 or c(c1, s2)

#> {1, (1L 2L), 2}

s2 + s1     # same

#> {2, (1L 2L), 1}

s1 - s2

#> {1}

s2 - s1

#> {2}

s1 = setnew(1, b = 2)
s2 = setnew(1, b = 4)
s1 & s2     # {1}

#> {1}

s1 | s2     # {1, b = 2, b = 4}

#> {1, b = 2, b = 4}